Grammartreeize drawing.c
Added noun_ND for an adverbed noun.

This change adds a very important TODO. Currently, object and monster symbol
names are grammartree-ized. They should not be.

It also adds some more features to check_intl: the engine now detects nointl
comments (as opposed to other comments) and !nointl{ and }nointl! can be used
inside comments to delimit large blocks of code.

(It occurs to me that just using a macro #defined to nothing might be
another option, with BEGIN_NOINTL and END_NOINTL. But we already use comments so
I won't buck that trend)