Work in progress, backing it up to a new branch
The idea is to automatically generate much of the repetitive code
for data structure manipulation that currently has to be manually
maintained. This stuff is all boilerplate, and there are files full
of it, which often end up with mistakes in it.

This work is far from finished, but I wanted a backup of it, and the
repo is a good place for that.