Get the structdesc parser working
types.sd and structdesc.sd now parse correctly into the compiled
internal format. Presently, there is only as much error checking
as is necessary for a correct parse (the rest will eventually be
done by checking against the description in structdesc.sd), and
nothing can process the format yet. But it's a start!