A minor optimization in directed chamber search
This has brought the time down for a remcap 3 directed chamber
from "so long I stopped waiting" to "several minutes". So far, it's
found two different such chambers:

    ---    ---
  ---4--  --4----
 --.0..|  |...0.|
--.0...|  |.0--.--
|.0..---  |..0.0.|
|...--    --.....|
--@--      --@----
 ---        ---

It's possible that there are others, but there can't be that many
within the search radius I'm using.

My current plan is either to never use this sort of directed chamber,
or else to hardcode a list of such chambers and use them very rarely
(so that players don't notice the repetition). Unless I'm missing
something major, it doesn't make sense to generate these live.