Bugfix when there are no valid layouts
This happens if, say, every square on which you could place a crate
locks you in. Example layout to trigger the bug:

-----
|...|
|.-.|
|...----
--.-...|
 |...0.|
 |@-----
 ---

This layout isn't rejected out of hand, because the crate shown
looks like a valid storage location. However, the player will get
locked in if they try to push the crate out, and all other locations
either collapse to that one, or are connectivity locked.

The fix is simply to reroll the chamber when this happens.